packages/destination-actions/src/destinations/acoustic-s3tc/receiveEvents/__tests__/sftpCache.test.ts
import { cache2SFTP } from '../sftpCache' // Assuming you export the functions
import sftp from 'ssh2-sftp-client'

jest.mock('@segment/actions-core')
jest.mock('ssh2-sftp-client')
jest.mock('path')
jest.mock('../generated-types')

const validSFTPSettings = {
  cacheType: 'SFTP',
  sftp_user: 'username',
  sftp_password: 'password',
  sftp_folder: '/path/to/folder',
  fileNamePrefix: 'prefix',
  __segment_internal_engage_force_full_sync: false,
  __segment_internal_engage_batch_sync: false
}

describe('cache2SFTP', () => {
  it('should expose a function', () => {
    expect(cache2SFTP).toBeDefined()
  })

  it('cache2SFTP should return expected output', async () => {
    const retValue = await cache2SFTP(new sftp(), validSFTPSettings, 'file_name_is_this.csv', 'fileContent')
    expect(retValue).toBeUndefined()
  })
})

describe('cache2SFTP', () => {
  test('uploads file content to SFTP server', async () => {
    // Mock the necessary dependencies and perform assertions
  })

  // Add more test cases for different scenarios and edge cases
})

describe('testAuthSFTP', () => {
  test('checks SFTP authentication by verifying folder existence', async () => {})
})
